feat: Add formatNumber utility and apply price and high/low value formatting in TokenList
